Suppress warnings about signed/unsigned comparisons.
CLANG enables this warnings by default in C mode.

+++ src/defs.c
@@ -79,7 +79,7 @@ static int case_name_cmp_str(const char *a, const char *b)
 
 int defs_search(const defs *d, const char *name)
 {
-    int i;
+    size_t i;
     for(i=0; i<darray_len(d); i++)
         if( !case_name_cmp(name, darray_i(d,i).name) )
             return i;
@@ -120,7 +120,7 @@ int defs_new_def(defs *d, const char *name, const char *file_name, int file_line
     return i;
 }
 
-void defs_set_string(defs *d, int id, const char *data, int len)
+void defs_set_string(defs *d, unsigned id, const char *data, int len)
 {
     assert( id >= 0 && id < darray_len(d) );
     darray_i(d,id).data = malloc(len);
@@ -128,32 +128,32 @@ void defs_set_string(defs *d, int id, const char *data, int len)
     darray_i(d,id).len = len;
 }
 
-void defs_set_numeric(defs *d, int id, double val)
+void defs_set_numeric(defs *d, unsigned id, double val)
 {
     assert( id >= 0 && id < darray_len(d) );
     darray_i(d,id).val = val;
 }
 
-const char *defs_get_string(const defs *d, int id, int *len)
+const char *defs_get_string(const defs *d, unsigned id, int *len)
 {
     assert( id >= 0 && id < darray_len(d) && darray_i(d,id).data );
     *len  = darray_i(d,id).len;
     return darray_i(d,id).data;
 }
 
-double defs_get_numeric(const defs *d, int id)
+double defs_get_numeric(const defs *d, unsigned id)
 {
     assert( id >= 0 && id < darray_len(d) && !darray_i(d,id).data );
     return darray_i(d,id).val;
 }
 
-int defs_get_type(const defs *d, int id)
+int defs_get_type(const defs *d, unsigned id)
 {
     assert( id >= 0 && id < darray_len(d) );
     return darray_i(d,id).data != 0;
 }
 
-const char * defs_get_name(const defs *d, int id)
+const char * defs_get_name(const defs *d, unsigned id)
 {
     assert( id >= 0 && id < darray_len(d) );
     return darray_i(d,id).name;
